Final fixes to make things actually work
