+++ .github/workflows/todo.yaml
@@ -1,0 +1,11 @@
+name: ci
+
+on: push
+jobs:
+  todo:
+    runs-on: "ubuntu-latest"
+    steps:
+      - uses: "actions/checkout@master"
+      - name: "TODO to Issue"
+        uses: "alstr/todo-to-issue-action@v4.6.6"
+        id: "todo"
