+++ pkg/builtin_plugins/gifs/gifs.go
