+++ pkg/plugin_manager/reactions.go
@@ -1,14 +1,15 @@
 package plugin_manager
 
-import "github.com/slack-go/slack/slackevents"
+import (
+	"github.com/slack-go/slack/slackevents"
+)
 
 // dispatchReactions sends a reaction to all registered reaction hooks
 func (m *ManagerImpl) dispatchReactions(ev *slackevents.ReactionAddedEvent) {
 	for _, rh := range m.reactionHooks {
 		rh.ReactionHook.Channel() <- &ReactionHookMsg{
-			Bot:      b,
+			Helper:   NewPluginHelper(rh.PluginID, m.l, m.slackManager, m.dataStore.GetStore(rh.PluginID)),
 			Reaction: ev,
-			Store:    b.getStore(rh.PluginId),
 		}
 	}
 }
