+++ pkg/slack_client/wire.go
@@ -1,0 +1,11 @@
+package slack_client
+
+import (
+	"github.com/google/wire"
+)
+
+var Wired = wire.NewSet(
+	NewConfig,
+	wire.Struct(new(SlackHttpClient)),
+	NewSlackClient,
+)
