--- pkg/slack_manager/client/wire.go
@@ -1,11 +1,0 @@
-package client
-
-import (
-	"github.com/google/wire"
-)
-
-var Wired = wire.NewSet(
-	NewConfig,
-	wire.Struct(new(slackHttpClient)),
-	NewSlackClient,
-)
